requirements.txt
attrs==19.3.0
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cryptography==3.0
Fiona==1.8.13
Flask==1.1.2
geopandas==0.8.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.1
munch==2.5.0
numpy==1.19.1
olefile==0.46
pandas==1.1.0
Pillow==7.2.0
pip==20.2.2
psycopg2==2.8.5
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.6.1.post1
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
requests==2.24.0
Rtree-linux==0.9.4
Shapely==1.7.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
Werkzeug==1.0.1
wheel==0.35.1
win-inet-pton==1.1.0
wincertstore==0.2
